refactor(hickey): extract <CodeMenuFrame> for menu wiring
CodeTab.tsx (FileDiff render) and BrowseFileView.tsx (FileView render)
both wrapped a Pierre viewer in the same six-line scaffold: host
<div> + addEventListener('contextmenu') + useLineSelection +
CodeContextMenu. "Wire kolu's path:line menu onto a Pierre viewer"
is a single concept; lifting it into <CodeMenuFrame> removes the
structural repetition without altering either call site's behavior.
